Move 404 to domains, drop shared
