+++ templates/hotel-booking.config.js
@@ -1,6 +1,7 @@
-import { header, summary, notFound, footer } from './hotel-booking/layout/index.js';
+import { header, summary, footer } from './hotel-booking/layout/index.js';
 import { checkout, confirmation, finalPriceConsent, rooms } from './hotel-booking/sections/index.js';
 import error from './hotel-booking/error.js';
+import notFound from './hotel-booking/not-found.js';
 
 export default {
     layout: {
@@ -9,9 +10,7 @@ export default {
         notFound,
         footer
     },
-    cache: [
-
-    ],
+    cache: [],
     pages: [
         {
             name: 'rooms',
