+++ templates/internal/layout/index.js
@@ -1,11 +1,9 @@
 import summary from './summary.js';
 import header from './header.js';
 import footer from './footer.js';
-import notFound from '../../shared/not-found.js';
 
 export {
     summary,
-    notFound,
     header,
     footer
 };
