+++ templates/motor-insurance.config.js
@@ -1,6 +1,7 @@
-import { header, summary, notFound, footer } from './motor-insurance/layout/index.js';
+import { header, summary, footer } from './motor-insurance/layout/index.js';
 import { quote, quoteLoading, options, assumptions, statements, feesSummary, declarations, marketingConsent, summaryPage, checkout, finalPriceConsent, confirmation } from './motor-insurance/sections/index.js';
 import error from './motor-insurance/error.js';
+import notFound from './motor-insurance/not-found.js';
 
 export default {
     layout: {
